move validation functions to helper.ts
